src/requirements.txt
asgiref==3.5.2
attrs==22.1.0
certifi==2022.12.7
chardet==4.0.0
Django==3.2
djangorestframework==3.14
ecdsa==0.18.0
httpie>=2.4.0
idna==3.4
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
Pygments==2.8.1
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.2.0
pytest-django==4.5.2
pytz==2022.6
requests==2.28.1
requests-toolbelt==0.10.1
six==1.16.0
sqlparse==0.4.3
toml==0.10.2
urllib3==1.26.13
Sphinx>=2.0.0
sphinxcontrib-napoleon>=0.7
sphinx-rtd-theme>=0.4.3
